+++ nuttx/configs/shenzhou/src/Makefile
@@ -35,87 +35,87 @@
 
 -include $(TOPDIR)/Make.defs
 
-CFLAGS		+= -I$(TOPDIR)/sched
+CFLAGS += -I$(TOPDIR)/sched
 
-ASRCS		=
-AOBJS		= $(ASRCS:.S=$(OBJEXT))
+ASRCS =
+AOBJS = $(ASRCS:.S=$(OBJEXT))
 
-CSRCS		= up_boot.c up_spi.c up_mmcsd.c up_chipid.c
+CSRCS = stm32_boot.c stm32_spi.c stm32_mmcsd.c stm32_chipid.c
 
 ifeq ($(CONFIG_HAVE_CXX),y)
-CSRCS		+= up_cxxinitialize.c
+CSRCS += stm32_cxxinitialize.c
 endif
 
 ifeq ($(CONFIG_ARCH_LEDS),y)
-CSRCS		+= up_autoleds.c
+CSRCS += stm32_autoleds.c
 else
-CSRCS		+= up_userleds.c
+CSRCS += stm32_userleds.c
 endif
 
 ifeq ($(CONFIG_ARCH_BUTTONS),y)
-CSRCS		+=  up_buttons.c
+CSRCS +=  stm32_buttons.c
 endif
 
 ifeq ($(CONFIG_ARCH_RELAYS),y)
-CSRCS		+=  up_relays.c
+CSRCS +=  stm32_relays.c
 endif
 
 ifeq ($(CONFIG_STM32_OTGFS),y)
-CSRCS		+= up_usb.c
+CSRCS += stm32_usb.c
 endif
 
 ifeq ($(CONFIG_NSH_ARCHINIT),y)
-CSRCS		+= up_nsh.c
+CSRCS += stm32_nsh.c
 endif
 
 ifeq ($(CONFIG_MTD_W25),y)
-CSRCS		+= up_w25.c
+CSRCS += stm32_w25.c
 endif
 
 ifeq ($(CONFIG_USBMSC),y)
-CSRCS		+= up_usbmsc.c
+CSRCS += stm32_usbmsc.c
 endif
 
 ifeq ($(CONFIG_USBDEV_COMPOSITE),y)
-CSRCS		+= up_composite.c
+CSRCS += stm32_composite.c
 endif
 
 ifeq ($(CONFIG_CAN),y)
-CSRCS		+= up_can.c
+CSRCS += stm32_can.c
 endif
 
 ifeq ($(CONFIG_ADC),y)
-CSRCS		+= up_adc.c
+CSRCS += stm32_adc.c
 endif
 
 ifeq ($(CONFIG_WATCHDOG),y)
-CSRCS		+= up_watchdog.c
+CSRCS += stm32_watchdog.c
 endif
 
 # NOTE: SSD1289 is not supported on the board
 
 ifeq ($(CONFIG_LCD_SSD1289),y)
-CSRCS		+= up_ssd1289.c
+CSRCS += stm32_ssd1289.c
 else
-CSRCS		+= up_ili93xx.c
+CSRCS += stm32_ili93xx.c
 endif
 
 ifeq ($(CONFIG_INPUT_ADS7843E),y)
-CSRCS		+= up_touchscreen.c
+CSRCS += stm32_touchscreen.c
 endif
 
-COBJS		= $(CSRCS:.c=$(OBJEXT))
+COBJS = $(CSRCS:.c=$(OBJEXT))
 
-SRCS		= $(ASRCS) $(CSRCS)
-OBJS		= $(AOBJS) $(COBJS)
+SRCS = $(ASRCS) $(CSRCS)
+OBJS = $(AOBJS) $(COBJS)
 
-ARCH_SRCDIR	= $(TOPDIR)/arch/$(CONFIG_ARCH)/src
+ARCH_SRCDIR = $(TOPDIR)/arch/$(CONFIG_ARCH)/src
 ifeq ($(WINTOOL),y)
-  CFLAGS	+= -I "${shell cygpath -w $(ARCH_SRCDIR)/chip}" \
-  		   -I "${shell cygpath -w $(ARCH_SRCDIR)/common}" \
-  		   -I "${shell cygpath -w $(ARCH_SRCDIR)/armv7-m}"
+  CFLAGS += -I "${shell cygpath -w $(ARCH_SRCDIR)/chip}" \
+            -I "${shell cygpath -w $(ARCH_SRCDIR)/common}" \
+            -I "${shell cygpath -w $(ARCH_SRCDIR)/armv7-m}"
 else
-  CFLAGS	+= -I$(ARCH_SRCDIR)/chip -I$(ARCH_SRCDIR)/common -I$(ARCH_SRCDIR)/armv7-m
+  CFLAGS += -I$(ARCH_SRCDIR)/chip -I$(ARCH_SRCDIR)/common -I$(ARCH_SRCDIR)/armv7-m
 endif
 
 all: libboard$(LIBEXT)
