+++ nuttx/configs/shenzhou/src/stm32_buttons.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * configs/shenzhou/src/board_buttons.c
+ * configs/shenzhou/src/stm32_buttons.c
  *
  *   Copyright (C) 2012, 2014-2015 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
