+++ nuttx/configs/shenzhou/src/stm32_chipid.c
@@ -1,7 +1,7 @@
 /************************************************************************************
- * configs/shenzhou/src/up_boot.c
+ * configs/shenzhou/src/stm32_chipid.c
  *
- *   Copyright (C) 2012, 2015 Gregory Nutt. All rights reserved.
+ *   Copyright (C) 2012 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
  *
  * Redistribution and use in source and binary forms, with or without
@@ -39,13 +39,11 @@
 
 #include <nuttx/config.h>
 
-#include <debug.h>
+#include <stdio.h>
 
-#include <nuttx/board.h>
 #include <arch/board/board.h>
 
 #include "up_arch.h"
-#include "shenzhou-internal.h"
 
 /************************************************************************************
  * Pre-processor Definitions
@@ -59,44 +57,35 @@
  * Public Functions
  ************************************************************************************/
 
-/************************************************************************************
- * Name: stm32_boardinitialize
- *
- * Description:
- *   All STM32 architectures must provide the following entry point.  This entry point
- *   is called early in the intitialization -- after all memory has been configured
- *   and mapped but before any devices have been initialized.
- *
- ************************************************************************************/
-
-void stm32_boardinitialize(void)
+const char *stm32_getchipid(void)
 {
-  /* Configure SPI chip selects if 1) SPI is not disabled, and 2) the weak function
-   * stm32_spiinitialize() has been brought into the link.
-   */
+  static char cpuid[12];
+  int i;
 
-#if defined(CONFIG_STM32_SPI1) || defined(CONFIG_STM32_SPI3)
-  if (stm32_spiinitialize)
+  for (i = 0; i < 12; i++)
     {
-      stm32_spiinitialize();
+      cpuid[i] = getreg8(0x1ffff7e8+i);
     }
-#endif
 
-  /* Initialize USB is 1) USBDEV is selected, 2) the USB controller is not
-   * disabled, and 3) the weak function stm32_usbinitialize() has been brought
-   * into the build.
-   */
+  return cpuid;
+}
 
-#if defined(CONFIG_USBDEV) && defined(CONFIG_STM32_USB)
-  if (stm32_usbinitialize)
+const char *stm32_getchipid_string(void)
+{
+  static char cpuid[27];
+  int c;
+  int i;
+
+  for (i = 0, c = 0; i < 12; i++)
     {
-      stm32_usbinitialize();
+      sprintf(&cpuid[c], "%02X", getreg8(0x1ffff7e8+11-i));
+      c += 2;
+      if (i % 4 == 3)
+        {
+          cpuid[c++] = '-';
+        }
     }
-#endif
 
-  /* Configure on-board LEDs if LED support has been selected. */
-
-#ifdef CONFIG_ARCH_LEDS
-  board_led_initialize();
-#endif
+  cpuid[26] = '\0';
+  return cpuid;
 }
