+++ nuttx/configs/shenzhou/src/stm32_relays.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * configs/shenzhou/src/up_relays.c
+ * configs/shenzhou/src/stm32_relays.c
  *
  *   Copyright (C) 2012 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
