+++ nuttx/configs/shenzhou/src/stm32_touchscreen.c
@@ -1,6 +1,5 @@
 /************************************************************************************
- * configs/shenzhou/src/up_touchscreen.c
- * arch/arm/src/board/up_touchscreen.c
+ * configs/shenzhou/src/stm32_touchscreen.c
  *
  *   Copyright (C) 2011-2012 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
@@ -55,7 +54,7 @@
 #include "shenzhou-internal.h"
 
 /****************************************************************************
- * Pre-Processor Definitions
+ * Pre-processor Definitions
  ****************************************************************************/
 /* Configuration ************************************************************/
